fix(ROB-586): re-parent migration onto 885e50ac5bb1; drop foreign_holding dup
ROB-575 (merged to main) already adds foreign_holding_shares/foreign_holding_rate
via 20260615_rob575_investor_flow_snapshot_market_fields, so this migration now
only adds discussion_sentiment_rank and chains onto the current head (resolves
the 2-head alembic conflict that failed test_revision_graph_has_single_final_head).
